move prepare_warming_level
